requirements.txt
# Feel free to add your Python dependencies below. For formatting guidelines, see:
# https://pip.pypa.io/en/latest/reference/requirements-file-format/

apify < 4.0.0
apify-client < 3.0.0
fastmcp>=2.14.0
mcp>=1.25.0
python-dotenv